// apps/web/src/lib/service-clients/service-properties/generated/schemas/entityReference.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * properties_service
 * OpenAPI spec version: 0.1.0
 */

import type { EntityReferenceSpecificMessageId } from './entityReferenceSpecificMessageId';
import type { EntityType } from './entityType';

/**
 * Entity reference for entity-type property values.
 */
export interface EntityReference {
  entity_id: string;
  entity_type: EntityType;
  /** For CHANNEL, CHAT, THREAD entity types - optional specific message ID.
This allows referencing a specific message within a thread/channel/chat. */
  specific_message_id?: EntityReferenceSpecificMessageId;
}